[nwp] Fixed accumulation of small epsilon on contravariant mass flux (in PPM)
See merge request icon/icon-nwp!1455

In the Piecewise Parabolic Method (PPM) subroutine, when the `l_out_edgeval` option is set to true, a small epsilon is added to the elements of the contravariant mass flux (`p_mflx_contra_v`). This is done in order to avoid any division by zero in a later expression where we divide by `p_mflx_contra_v`. However, when adding the epsilon to an element of the array we do not check that such element is equal to zero, but we add it to all elements indiscriminately. In principle this is fine, since epsilon is only a small value, however we need to take into account the fact that the PPM is called for every tracer and we add an epsilon for every one of those. This may be a problem for simulations using a large number of tracers, since the small values quickly accumulate on `p_mflx_contra_v`. Moreover, this also introduces an unnecessary dependency between tracers that blocks any attempt to parallelize over the tracer dimension (assuming one wants to have bit-identical results).

Here we propose a solution for this problem, where we only add an epsilon to the denominator of the division by `p_mflx_contra_v` without updating the field itself.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>