[nwp] Fix to use the right pointer to and index of tracers in the generalized meteogram output
See merge request icon/icon-nwp!1461

Fix to use the right pointer to and index of tracers in the generalized meteogram output.

While this feature is mainly used to include ART variables (tracers and diagnostics) in the meteogram output, it is nevertheless independant of ART.
I. e. there is no related change in an ART branch.

Approved-by: Liselotte Bach <liselotte.bach@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>