[mpim] Expose BUILD_ENV to the runscript generators
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/496

This MR makes it possible to run the BUILD_ENV script, normally set in a
configure wrapper, as part of a runscript. This feature is discouraged as it is
generally a bad idea to implement configure wrappers that produce executables
that require the build environment at the runtime. However, it seems to be a
lesser evil than duplicating the module loading in the runscripts.

Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>