Preparing a major revision of turbulence code including the integration of some not yet considered effects of surface roughness
See merge request icon/icon-nwp!554

By this branch, referred to as 'mrprep', new contributions from the COSMO action ConSAT (aiming at "Consolidating the general Surface-to-Atmosphere Transfer" (and related inter actions of ICON components) by means of the modules 'turb_diffusion', 'turb_transfer' and 'sfc_terra' including related interfaces and utility routines) are prepared for being stepwise merged into the 'master'. 

For the current mrege-request, it contains a **first step** of the first package. 

The whole first package covers so far missing effects of surface roughness in the atmospheric part of ICON. These additional effects are mainly the consideration of (a) subgrid-scale thermal circulations related to SSO (katabatic and anabatic circulations), (b) additional roughness by tile-variation of land use, (c) additional surface-shear by non-turbulent subgrid-scale circulations (NTCs) and (d) snow-sinking of land-use roughness. Further, (d) deposition-fluxes of cloud-water/ice are now considered in 'terra', and (e) the diffusion of non-gradient fluxes (by NTCs or by corrections of turbulent scalar fluxes with regard to local saturation adjustment) has been reformulated. 
Moreover, (f) a stronger modularization and various measures of code optimization and cleaning, finally lead to an appreciable speed-up of the entire turbulence code and allows to get rid of some obsolete code.

Although all these isses are already present, and the full code has even passed all BB-tests in a neutral configuration (at which all new effects are switched off), this full code has temporaily been reduced again by code related to (a) and (b) in order to provide a **fist step** fitting better to our gate-keeping process. 

This remaining code conatains all the further generalizations and reformulations, which mainly are intricately connected. They also serve as a necessary preparation for the **second step** dealing with (a) and (b), which contains a major component of the whole develoment. This **first step** particularly contains (f), and it keeps the amount of code quite similar to that of the 'master'.

Before merging this **first step** into the 'master', The various small issues related to "ifdef OLD_CODE_EMUL" should be resolved by inspection together with the gate-keepers, so as to finally remove all "OLD_CODE_EMUL"ations.  

Before introducing the **second step**, some formal or technical interim-steps are planned as a measure of further consolidation of the turbulence-code: (i) Sorting out of obsolete options; (ii) Clearing the use of former/similar version of turbulent saturation-adjustment in cloud-diagnostics; (iii) Removing multiple calculation of derived surface-parameters; (iv) Translation of German comments into English; (v) Transition 'gz0'->'z0'; ...

Some other desired reorganization like, e.g., “getting rid of separate CALLs of SUB ‘turbran’ for tiled and not-tiled runs in ‘mo_nwp_turbtrans_interface’” or “straightening the initialization of turbulence with respect to tiles” as well as “arranging initialization of turbulence possibly by means of particular ‘_init’-routines” are strongly involved with the developments (a) and (b). Therefore, in order to avoid another grueling period of adapting an almost complete development to a modifying infrastructure, this kind of reorganization ought to be done either **after** activation of (a) and (b) in the master (which is here called the **second step**) or **within** this branch ‘mrprep’ (containing already the whole development, but with (a) and (b) “deactivated”). 

'mrprep' is also foreseen to integrate the 2-nd portion of new ConSAT contributions, which are more related to the non-atmospheric part of ICON, as it is represented by module 'srf_terra' (implicit treatment of surface processes including the implicit coupling of multi-layer fractional snow, as well as necessary adaptations in various parts of associated code), which are going to follow in near future.
This development is so far contained in branch 'mrsurf', which is being kept up to date with 'mrprep'.

@b381022

Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>