[nwp] mpi: check worker architecture during communicator creation
See merge request icon/icon-nwp!1454

A common user error on NEC Aurora is a mismatch between the process
layout of vector host and vector engine processes specified to
mpirun and to ICON. At the moment, such errors usually cause an MPI
error in mo_nh_diffusion because VEs exchange velocity fields in
double precision even when mixed precision is enabled while VH PEs
expect single-precision data.

To make this error more obvious and easy to fix, this MR checks
the architecture of all worker PEs that are not detached stdio PEs
(enabled by nonzero proc0_shift). A mismatch causes an abort with
an appropriate error message. A mismatch of the I/O PEs and the
detached stdio PEs is already detected in the same place.

Detection is achieved by setting different bits in an integer for
VH and VE PEs (currently 1 and 2), and performing a binary-or
reduce over all PEs in the p_comm_work_only communicator. If more
than a single bit is set in the result, multiple architectures are
present. The check is restricted to regular worker PEs. Removing
that restriction would force all PE groups to be homogeneous as
p_comm_work_only aliases the intra-group communicator p_comm_work
for all other PE groups.

This commit correctly sets the detached stdio PE count to
proc0_shift during ocean initialization. This has the side effect
that timers are now shown separately for the detached and regular
ocean workers.

Approved-by: Florian Prill <florian.prill@dwd.de>
Merged-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>