Extension of adaptive parameter tuning
See merge request icon/icon-nwp!1494

Next evolution step of adaptive parameter tuning (APT), including
- hydraulic conductivity of the upper two soil layer interfaces (complementing the APT for the bare-soil evaporation resistance)
- albedo over snow-free land (complementing the existing albedo adaptation over snow-covered land)
- diagnosis of snow-cover fraction, potentially relevant for snow depths less than 15 cm.

The new APT components come along with two additional predictor fields, containing filtered assimilation increments for temperature and relative humidity with a temporal weighting peaking in the afternoon (16:30 LT), which approximately corresponds to the time at which the maximum 2-m temperature is reached during the summer months.

In addition, a tuning parameter for the number of random-number spinup calls for the EPS perturbations is added because it became apparent during the testing phase that the EPS scores are a bit worse with gcc11 than with gcc9 when the current default of spinup calls is used.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>