[mpim] ICON-Land: Reduce output from water balance checks
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/514

Simulations that do not want to monitor or debug water conservation on the land surface could set
`enforce_water_budget = .FALSE.` to avoid the finish statement in the water balance check. However,
the water balance diagnostic was still written to the log file, and in case a large number of grid
cells were affected this caused the log file to grow in size.

Therefore, the logical `enforce_water_budget` switch is changed to a string (for the namelist,
internally an integer is used) to allow for more fine tuning of the water balance check. It now
offers the options `error` which causes the model to stop if a water balance problem is encountered
(similar to the former `enforce_water_budget == .TRUE.`), `logging` which prints the water balance
diagnostic and related messages to the log file (similar to the former `enforce_water_budget == .FALSE.`)
and `ignore` where the water balance diagnostic is not even written into the log file. The latter still
writes a single line warning if an problem is encountered.

Furthermore, all other soil hydrology sanity checks are now also connected to this namelist switch. As
with `enforce_water_budget`, it can be set either in the `jsb_model_nml` for both processes, hydro and hd,
or individually in the hydro and hd configs.

This commit only changes the diagnostic output. Model results are identical.

See https://gitlab.dkrz.de/jsbach/jsbach/-/merge_requests/219

Co-authored-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>
Approved-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>
Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>