[mpim] Restructuring and cleaning up in TMX
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/472

The type-bound procedure `Compute` of type `t_vdf` is restructured and cleaned up, containing
new subroutines for the turbulent diffusion of hydrometeors, temperature, horizontal and
vertical velocity, respectively. 

The type-bound procedure `Compute_diagnostics` of type `t_vdf_atmo` is restructured as a
high-level routine to manage all subroutine calls for the computation of the eddy viscosity
and diffusivity. 

A new module `mo_tmx_smagorinsky` contains the actual computation of the eddy viscosity and
diffusivity, and introduces a simple way to select different stability correction functions
without compromising readability. The turbulent mixing length calculation is moved to this module.

Co-authored-by: Kay Schäfer <kay.schaefer@uni-hamburg.de>
Co-authored-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>
Merged-by: Reiner Schnur <reiner.schnur@mpimet.mpg.de>