[mpim] [m300083-fix-runoff-diag] do runoff diagnostic only when received from yac
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/516

Currently, the runoff diagnostics is done within each ocean timestep even when nothing has been received from YAC due to the set coupling interval. This branch moves the diagnostics inside an already existing if-clause to ensure that diagnostics is only done when new data are available. As this diagnostics invokes a global sum I'd prefer to call it only when absolutely necessary. In particular for our high resolution coupled configurations I'd like to restrict global communication to the absolute necessary. One could also think about increasing the msg level for this diagnostics. The ocean already prints out a global sum of the runoff elsewhere. Thus we may not need this duplication other than for debugging purposes which would justify to increase the msg level to e.g. 15 or 20 for this additional diagnostics.

Merged-by: Monika Esch <monika.esch@mpimet.mpg.de>