[mpim] Interface aes ocean: move coupling step from mo_interface_iconam_aes.f90 to mo_nh_stepping.f90
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/509

The move of the coupling call from mo_interface_iconam_aes.f90 to mo_nh_stepping.f90 is necessary to accommodate the coupling of the nests (see milestone). The coupler would require the information from all the nested domains in order to choose and merge the higher resolution data. Under the mo_interface_iconam_aes.f90 the call is performed for each domain, in each sub-step, where such global information is not available. So it is moved outside the recursive, sub-stepping call for each domain, into the time-stepping. 

The results are binary identical to the previous implementation.

Co-authored-by: Leonidas Linardakis 
Co-authored-by: Moritz Hanke <hanke@dkrz.de>
Approved-by: Marco Giorgetta <marco.giorgetta@mpimet.mpg.de>
Merged-by: Monika Esch <monika.esch@mpimet.mpg.de>