Restructure 2mom microphysics scheme for 1mom granule and gscp=3
See merge request icon/icon-nwp!1527

## Feature Description

Restructuring of 2mom microphysics to prepare an implementation of the 2mom cloud ice scheme that is compatible with the 1mom granule. The restructuring introduces a new module mo_2mom_mcrph_setup, which contains the setup subroutines for the process parameterizations and the functions for the necessary coefficients. Some more technical subroutines have been moved from mo_2mom_mcrph_processes to mo_2mom_mcrph_util. There is no change in content.

This is a first part of the former !1479.

Approved-by: Alberto de Lozar <alberto.lozar-de@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>