[nwp] name_list_output: fix deadlock when meteogram output is used in a coupled simulation
See merge request icon/icon-nwp!1539

Meteogram setup is done before initializing the coupler on worker PEs but after on
IO PEs. The worker PE with a meteogram cell tries to send metadata to the IO PE
responsible for output while that IO PE waits for the worker to enter the coupler
initialization. The simulation deadlocks.

This MR moves the meteogram setup on IO PEs ahead of the coupler initialization.

Approved-by: Florian Prill <florian.prill@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>