Add boundary interpolation for ICON-D05 diagnostic fields
See merge request icon/icon-nwp!1577

The lateral boundary interpolation for nested domains is extended to a set of variables that is needed for a correct output in ICON-D05:
- precipitation / convection diagnostics computed in the operational ICON-D2 forecasts
- slope-dependent radiation fields
- 10-minute averaged 10m winds

This comes along with a bug fix for the position of the `fill_nestlatbc_phys` call, which needed to be moved after the `pp_scheduler` in order to execute the computations in the right order.

In addition, the interface of `interpol_scal_grf` has been modified in order to avoid further manual adjustments of some local array dimensions, and the execute permission for the ftrace config wrapper is restored.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>