[mpim] Optimization of simple plumes: Calculate geometric grid information only once in advance.
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/547

Optimize the calculation of the simple plumes by separating the calculation of the geometric information only once. The geometric grid information for the plumes requires the calculation of many cos,sin,exp and merges of arrays. This slows down the calculation. When four time steps are performed, radiation every second time step, about 10% of total computer time is saved with this method (the radiation is calculated once including geometric information, once without) in r2b4 resolution.

Co-authored-by: Sebastian Rast <m218036@levante2.lvt.dkrz.de>
Co-authored-by: Ali Sedighi <k202194@levante0.lvt.dkrz.de>
Co-authored-by: Sebastian Rast <m218036@levante0.lvt.dkrz.de>
Co-authored-by: Ali Sedighi <k202194@levante3.lvt.dkrz.de>
Co-authored-by: Ali Sedighi <k202194@levante4.lvt.dkrz.de>
Co-authored-by: Ali Sedighi <k202194@levante1.lvt.dkrz.de>
Merged-by: Sebastian Rast <sebastian.rast@mpimet.mpg.de>