Major update of two-moment cloud ice scheme (inwp_gscp=3)
See merge request icon/icon-nwp!1556

## Feature Description

This is a major revision of the two-moment cloud ice scheme for the global ICON model. To achieve good results in the tropics as well as in mid-latitudes, an explicit dependency on latitude is introduced. This can be motivated by the different properties of convectively generated cirrus in the tropics (often dominated by freezing of cloud droplets) vs large-scale frontal cirrus in mid-latitudes (heterogeneous ice nucleation in large-scale ascent). In practise, the latitude dependency was in fact necessary due to issues with the temperature profile in the tropics.

Contentwise, this is the first part of the former !1479. 

This MR fixes a minor issue of !1527.

Approved-by: Alberto de Lozar <alberto.lozar-de@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>