[nwp] Effective radius and snow drift for cloudice2mom (gscp3)
See merge request icon/icon-nwp!1581

This adds the consistent treatment of effective radius for the two-moment cloud ice scheme (gscp=3). It introduces a separate effective radius subroutine which has a different treatment of grid-scale vs sub-grid contributions. The latter helps to reduce the OLR bias in the global ICON when running cloudice2mom.

In addition, the cloud ice number source for the snow drift term is introduced.

This is the 2nd part of the former !1479. The first part was !1556.

Approved-by: Alberto de Lozar <alberto.lozar-de@dwd.de>
Merged-by: Daniel Rieger <377-m300368@users.noreply.gitlab.dkrz.de>