[nwp] Fixes to enable reproducibilty of ICON-CLM on GPU
See merge request icon/icon-nwp!1569

ICON_CLM was running on GPU, but not giving reproducible results (with
restarts or even after starting the same experiment twice). The reason
was the soil-moisture dependent albedo tuning (`tune_albedo_wso`).

The underlying reason is still somewhat unclear, maybe it is GPU
gangs overtaking each other while processing the tile loop. But
the albedo correction should be tile-dependent because soil moisture
is. The current implementation keeps on overwriting the correction
so that the tile with the highest tile index on a given cell wins.
Its albedo is then used to compute direct albedo effects for all
tiles. We think that the last write to the correction might depend
on the run-time scheduling of the GPU gangs, causing the
nondeterministic behavior. We considered adding a `GANG(STATIC: 1)`
clause for deterministic scheduling but dismissed it because the
physically correct solution is only slightly more complicated.

This MR also fixes a bug in the O3 interface for `irad_o3=5` where
a temporary array was not properly made OpenMP private. The changed
ICON_CLM test uses `irad_o3=5`, which lead to failures on hybrid
builders.

Approved-by: Marek Jacob <1129-b380572@users.noreply.gitlab.dkrz.de>
Merged-by: Daniel Rieger <377-m300368@users.noreply.gitlab.dkrz.de>