[nwp] dkrz aurora config: switch ocean to vectorized version
See merge request icon/icon-nwp!1597

Switch ocean on levante_aurora to vectorized version. This
works around a compiler bug triggered by the recent removal
of `!NEC$ options "-O0"` from `mo_ocean_tke.f90`.

The bug is caused by implicit copy-ins generated for
`calculate_density_onColumn`, which expects contiguous
arrays for temperature and salinity but is passed depth
slices for a single column.

The vectorized version flips the loops such that the cell
loop is the inner loop and calls the scalar version of the
routine, so no copy-in is needed. It also has massively
improved performance on vector engines.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>