[nwp] atmo_ocean_coupling: register ocean velocity fields with YAC only when vdiff is in use
See merge request icon/icon-nwp!1591

## Feature Description
Ocean velocity fields are not retrieved by the atmosphere with other
turbulence schemes than vdiff. If coupling.yaml defines a coupling
pair for the velocities in these cases, MPI requests pile up that
never complete. This leads to strange effects from linear slowdowns
in the number of coupling steps to outright deadlocks.

This change conditions the field registration during coupling setup
on the turbulence scheme. If the user tries to define a coupling pair
when vdiff is not used, YAC will throw an appropriate error message
and abort the model.

Approved-by: Moritz Hanke <hanke@dkrz.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>