[nwp] Fix initialization of transport-only tests
See merge request icon/icon-nwp!1600

The observed convergence rate of the horizontal transport schemes is significantly affected by the way the velocity field is specified.
This has been observed for the deformational flow test case and is potentially valid for any other testcase which relies on an analytic, time dependent velocity field.

In order to achieve the correct convergence rates (i.e. convergence rates which are expected based on theoretical considerations), it turned out to be essential to evaluate the time dependent velocity field at time $`t^{n+1/2}`$. In ICON, this was implemented with the following second order accurate (linear) approximation:
```math
v^{n+1/2} = 0.5 (v^{n} + v^{n+1})
```
This is basically equivalent to the approximation used inside the dynamical core when computing time-averaged velocities and mass fluxes for tracer transport. It turned out, however, that this approximation degrades the convergence rates at high resolutions, when the temporal discretization error becomes comparable in size to the spatial discretization error.

With this MR, the second order accurate approximation presented above is replaced by an exact evaluation of the analytic velocity field at $`t^{n+1/2}`$.


#### Results

**Deformational flow test with Gaussian hill initial condition**

* CFL_max 0.23
* no limiter
* linear reconstruction

**Corrected implementation in blue**

| $`L_{2}`$ error | $`L_{\infty}`$ error |
| ------ | ------ |
|  ![convergence_linear_c0228_Q7_l2_fix_init](/uploads/bf331acea3a5ccfdbd6fc43e1458a855/convergence_linear_c0228_Q7_l2_fix_init.png)      | ![convergence_linear_c0228_Q7_linfty_fix_init](/uploads/468bd39af82a5294eecd0e69b46f9dad/convergence_linear_c0228_Q7_linfty_fix_init.png)     |

Approved-by: Guenther Zaengl <guenther.zaengl@dwd.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>