Option for more accurate coefficients for saturation pressure
See merge request icon/icon-nwp!1609

The coefficients used for computing the saturation vapor pressure in ICON-NWP are inaccurate at temperatures below about -50°C, particularly with respect to water. As a consequence, the RH w.r.t. water at ice saturation is inaccurate as well.

Here, a new option is introduced (`itype_satpres_coeffs` in the NWP physics namelist) that allows switching to the more accurate coefficients of the IFS. This also removes an inconsistency between the parameterizations inherited from the COSMO model and the IFS, respectively, because the latter already use the more accurate IFS coefficients.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>