Tuning option to reduce activity of grayzone deep convection, to be used in ICON-D2
See merge request icon/icon-nwp!1614

This MR introduces a new tuning option for the grayzone deep convection used in ICON-D2, which sets an offset in the CAPE closure in order to further reduce the activity of the convection scheme. This is supposed to get into the next parallel routine cycle together with the graupel / satad retuning.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>