[nwp] Update visibility diagnostic to be consistent with `icpl_rad_reff=1` (typical RUC settings)
See merge request icon/icon-nwp!1604

The visibility diagnostic is updated to be consistent with typical RUC settings `icpl_rad_reff=1`, where the diagnostic `q_c` and `q_i` are modified to include a contribution of precipitating liquid and frozen particles, respectively. This modification is not consistent with the assumptions made in the VIS diagnostic. For that reason the modifications are locally undone in the diagnostic routine. Note that this is a provisional fix; a clean implementation would require additional storage arrays for the diagnostic QC and QI including precipitating hydrometeors.

In addition, the time level of the grid-scale moisture tracers passed to the visibility routine is corrected (which slightly changes the results in any case), and a tuning option for the day-night conversion factor is introduced (which significantly improves the scores for low visibilities at night).

Approved-by: Alberto de Lozar <alberto.lozar-de@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>