[mpim] Partially revert "[mpim] Replace math-support and math-interpolation with iconmath"
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/594

This partially reverts icon/icon-mpim!506. Namely, the MR disables the
'horizontal' component of the 'iconmath' library and restores the respective
code in the 'src/shr_horizontal' directory of ICON. For the reasons to be
investigated later, the reverted changes prevent us from synchronization with
the release candidate due to a failing tolerance test by the 'DWD_nec' builder.

Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>