[nwp] icon-dev.checksuite: fix MPI test and make nproma test more robust
See merge request icon/icon-nwp!1557

## Feature Description

Instead of fragile SED substitutions, provide explicit number of nodes and
MPI processes per node to the experiment scripts via environment variables
for the MPI test. The nproma test reads the base values of nproma or
nblocks_c from its environment and passes the modified variables to the
experiment via its environment.

### Background
The icon-dev checksuite operates in "russian doll" mode where the check
script is created by `make_target_runscript` and calls that same script to
create the experiment script itself. That script tacks a header onto the
check and experiment script that provides base values for parameters that
vary between machines and experiments, such as MPI process layout or nproma.
As such, parameters that get communicated to the batch system are fixed by
the check script, and their values should override the values in the
experiment script.

### Problem
Recently, in the wake of getting the checksuite working for the ICON-CLM
test (!1363), a first step was taken in that direction by propagating
the MPI process layout from the check to the experiment script. This,
however, broke the MPI test in multiple ways: First, the number of procs
could not be determined anymore because of a failed grep match. That was
fixed in icon/icon!423. Second, the check script relies on a SED
substitution to change the number of MPI procs that also no longer matches
any line in the experiment script. Thus, the MPI test silently runs with
the same number of processes as the reference run.

### Solution
Since the experiment script no longer hard-codes the MPI process layout,
we can actually pass modified environment variables to change the layout
for the test. The same is done for nproma, where we can replace the
parse-modify-substitute mechanism by a much simpler manipulation of the
`nproma` and `nblocks_c` environment variables.

The side effect of these changes is that nondefault parameters like those
specified in `create_target_header` must be specified for the check script
instead of the experiment script. Due to the naming convention that pairs
up `exp.<experiment>` with `check.<experiment>` and looking for
`*<experiment>*` to determine the parameters to apply, this works for all
current experiments. An additional benefit is that any flags passed to
`addexp` in the create_list scripts actually apply to the experiment
scripts as well as the check scripts.

Approved-by: Marek Jacob <1129-b380572@users.noreply.gitlab.dkrz.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>