Small correction for IFS saturation pressure coefficients
See merge request icon/icon-nwp!1627

Closer inspection revealed that the IFS saturation pressure functions are defined with respect to the triple point, whereas ICON uses the freezing point at atmospheric pressure as reference. This implies that the reference saturation pressure (c1es) needs to be the same (610.78 Pa) for both sets of coefficients.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>