Removed i_am_accel_node in all but a few modules
https://gitlab.dkrz.de/icon/icon/-/merge_requests/477

This is the result of the cleanup hackathon Dec. 2-4 with the objective of eliminating i_am_accel_node 
from ICON. Much of the code has been touched, but in general the changes are simple:

  * Added an argument `lacc` (generally mandatory) down through the call tree, which denotes the GPU context;
  * If running with a GPU-accelerator run on the GPU (.TRUE.) or the CPU (.FALSE.);  
  * If running on CPU-only, this argument will be ignored since it is only used in OpenACC directives;
  * At the highest level, pass `lacc=.TRUE.` (run on GPU) or `lacc=.FALSE.`, depending on the context.

Pipeline: https://gitlab.dkrz.de/icon/icon-cscs/-/pipelines/91185

Status: 32 occurrences of i_am_accel_node remaining in 5 ICON modules. We could not finalize the removal
because there are still occurrences within externals/jsbach, which require the participation of Reiner (to be addressed in January).

Co-authored-by: Dominik Zobel <zobel@dkrz.de>
Co-authored-by: Marek Jacob <1129-b380572@users.noreply.gitlab.dkrz.de>
Co-authored-by: Pradipta Samanta <samanta@dkrz.de>
Co-authored-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>
Co-authored-by: jonasjucker <jonas.jucker@c2sm.ethz.ch>
Co-authored-by: Christoph Müller <Christoph.Mueller@meteoswiss.ch>
Merged-by: William Sawyer <wsawyer@cscs.ch>