[mpim] Single precision extensions to mo_mpi
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/570

This introduces new p_pack and p_unpack polymorphic interfaces in the mo_mpi
module for the respective p_pack_* and p_unpack_* subroutines, which are
currently used in the code directly. The MR also extends the existing set of
functions with overloads for the single-precision type (mere copy/paste of the
double-precision subroutines).

Approved-by: Moritz Hanke <hanke@dkrz.de>
Merged-by: Sergey Kosukhin <sergey.kosukhin@mpimet.mpg.de>