[nwp] Fix a race condition in keeplatestn
See merge request icon/icon-nwp!1635

If `keeplatestn` could not find a specified file or directory, the
program crashed with an uncatched FileNotFoundError exception raised
from the `os.path.getmtime` call in `sorted`. This situation can be the
result of a race condition when `keeplatestn` is called with a glob on
the command line. Another process could remove files or directories
between the extension of the glob pattern on shell level and the
`os.path.getmtime` call in the code. The new solution implemented here
catches this case and excludes non-existent files from N.

Approved-by: Marek Jacob <1129-b380572@users.noreply.gitlab.dkrz.de>
Merged-by: Daniel Hupp <daniel.hupp@meteoswiss.ch>