[nwp] read_interface: use allocatable arrays instead of returning pointers
See merge request icon/icon-nwp!1620

## Feature Description
### Background
The read_interface functions provide an interface that returns owning
pointers to users. Often, users forget to deallocate these pointers,
causing memory leaks. Oftentimes, the leaks only happen during
initialization but, e.g., the bc_ozone module leaks memory during the
time loop.

### Solution
Fortran provides the means to express ownership over a chunk of heap-
allocated memory: the ALLOCATABLE attribute. There is no downside
compared to a POINTER. Ownership can be transferred between variables
via the MOVE_ALLOC intrinsic function, and non-owning pointers to
that chunk of memory can be obtained by giving the ALLOCATABLE the
TARGET attribute. However, it is impossible to leak memory because
the chunk gets deallocated once the ALLOCATABLE goes out of scope.

### Implementation
All read_interface routines have their `return_pointer` arguments
replaced by an `alloc_array` INTENT(OUT) argument that, on return,
contains an appropriately sized array filled with the data read from
file. The `alloc_array` and `fill_array` arguments are mutually
exclusive -- a fact that unfortunately can only be enforced at
runtime, because plain and ALLOCATABLE dummy arguments cannot be
distinguished during overload resolution.

All FUNCTIONs from read_netcdf_broadcast_2 are converted to
SUBROUTINEs with an additional `alloc_array` argument with the
same semantics as above. The returned pointer wouldn't have been
useful and even dangerous if the passed argument for `fill_array`
or `alloc_array` is not a TARGET (the returned pointer is
undefined in that case).

`read_1D`, `read_2D_time`, and `read_3D_time` retain the
`return_pointer` argument for compatibility with the ART submodule
for now. util_jsbach's routine have unchanged interfaces for
compatibility with the jsbach submodule. This is achieved by
performing a sourced allocation on the return pointer with a
temporary allocatable as source, and therefore involves a copy of
the data. These deprecated uses will be removed once the submodules
are updated. Incidentally, that update will remove a bunch of
memory leaks.

New types `t_alloc_2d`, etc., are provided in read_netcdf_types to
allow for arrays of ALLOCATABLEs in the same manner as `t_ptr_2d`
allow for arrays of POINTERs. These should be moved to the
fortran-support submodule.

As a drive-by fix, a missing deallocation of an mtime pointer is
added to mo_util_jsbach:is_time_ltrig_rad_m1. If only mtime would
separate its allocation and initialization functions, then we could
stack-allocate these structures... the ALLOCATABLE route is
impossible as well because it, too, would require using a different
allocation function in the C library.

For unknown reasons, the outputs of c2sm_clm_r13b03_seaice have
moved out of tolerance on balfrin. Speculation is that because an
allocatable is always contiguous and does not have any aliases,
the compiler optimizes differently to the POINTER case. The
nvfortran diagnostic output shows that, in the current master,
the compiler assumes that the target field and the pointer returned
by the read routine can alias. With the allocatable, these messages
disappear.

Approved-by: Moritz Hanke <hanke@dkrz.de>
Merged-by: Daniel Reinert <daniel.reinert@dwd.de>