[mpim] ICON-Land: Introduction of a variable group for land monitoring variables
https://gitlab.dkrz.de/icon/icon-mpim/-/merge_requests/639

The jsbach monitoring variables are now combined in the output group "jsb_mon".
With this, the monitoring variables no longer need to be listed individually in
the run scripts but can be referred to as a group.

For details check jsbach MR https://gitlab.dkrz.de/jsbach/jsbach/-/merge_requests/271

Approved-by: Karl-Hermann Wieners <karl-hermann.wieners@mpimet.mpg.de>
Merged-by: Veronika Gayler <veronika.gayler@mpimet.mpg.de>