Memory optimization for wave model
See merge request icon/icon-nwp!1668

To reduce the memory consumption of the wave model, the 4D fields sl and fl (wave source function and its derivative) are declared as `vp`, and the 4D fields for the direction- and frequency-dependent group velocity are removed at the expense of recomputing the directional projection at each time step. Taking both together, the runtime still decreases a little bit because using single precision speeds up the nonlinear wave interaction more than the cost of recomputing the group velocities. The reduction of the memory allocation is reduced by 10-30% depending on the model configuration.

Approved-by: Daniel Reinert <daniel.reinert@dwd.de>
Merged-by: Guenther Zaengl <guenther.zaengl@dwd.de>